Extend RobustGPSampler to allow sampling outside nominal ranges
Suppose the nominal range of parameter x is [L, U] and the noise range
is ±ε. The objective function may deteriorate sharply outside the
[L, U] range, but to learn this, it is necessary to be able to sample
values outside the [L, U] range.

To achieve this, the noisy_suggestion option is added, which allows
suggest_float() to return values with added noise rather than nominal
values.

In such cases, since the range for suggest_float() should be the range
after applying the disturbance, the nominal_ranges option is added to
specify nominal ranges separately.